Refactor: Create constants directory
* Add login file

* Add regex file
